requirements.txt
boto==2.48.0
django-storages==1.6.5
gunicorn==19.7.1
psycopg2>2.5.1
dj-static==0.0.6
juntagrico>=1.0.1
